backend/requirements.txt
# Production-ready FastAPI backend dependencies
# Security-hardened versions with known vulnerability patches

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# Input Validation
pydantic==2.5.0
email-validator==2.1.0

# Rate Limiting & Caching
redis==5.0.1
slowapi==0.1.9
python-redis-lock==4.0.0

# HTTP Client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Configuration & Environment
python-decouple==3.8
pydantic-settings==2.1.0

# Security Headers & CORS
fastapi-cors==0.0.6
secure==0.3.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing (dev/staging only)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Production Server
gunicorn==21.2.0